dashboard_react/backend/requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
